+++ opentelemetry-sdk/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Make `SpanProcessor.on_start` accept parent Context
   ([#1251](https://github.com/open-telemetry/opentelemetry-python/pull/1251))
+- Fix b3 propagator entrypoint
+  ([#1265](https://github.com/open-telemetry/opentelemetry-python/pull/1265))
 - Samplers to accept parent Context
   ([#1267](https://github.com/open-telemetry/opentelemetry-python/pull/1267))